requirements.txt
Django~=1.11.3
django-extensions~=1.5.7
djangorestframework~=3.2.3
dj-database-url~=0.4.1
gunicorn~=19.3.0
python-dotenv~=0.1.3
psycopg2~=2.7.0
